fix: #161 delete all data button is not correct style
