+++ client/src/components/web-extension/Common/DeleteAllDataButton/index.jsx
@@ -1,0 +1,52 @@
+import React, { useState } from 'react';
+import { useDispatch } from 'react-redux';
+import { useNavigate } from 'react-router-dom';
+import { Button } from 'react-bootstrap';
+import { ConfirmModal } from '@cd/components/Common/ConfirmModal';
+import { lockAccount } from '@cd/actions/userActions';
+
+const DeleteAllDataButton = () => {
+    const dispatch = useDispatch();
+    const navigate = useNavigate();
+	const [isOpenModal, setIsOpenModal] = useState(false);
+
+	const handleOnClick = () => {
+		setIsOpenModal(true);
+	};
+
+	const handleOnCloseModal = () => {
+		setIsOpenModal(false);
+	};
+
+    const handleOnOk = () => {
+        dispatch(lockAccount());
+        navigate('/connectAccount');
+    }
+
+	return (
+		<>
+			<Button variant="normal" onClick={handleOnClick}>Delete All Data</Button>
+			<ConfirmModal
+				isOpen={isOpenModal}
+				onOk={handleOnOk}
+				onClose={handleOnCloseModal}
+				title="Are you sure you want to delete your wallet ?"
+				description={
+					<div className="cd_setting_modal">
+						<div>
+							Your current wallet, accounts and assets will be removed from this app permanently. This
+							action cannot be undone.
+						</div>
+						<br />
+						<div>
+							Your can ONLY recover this wallet with your Secret Recovery Phase CasperDash does not have
+							your Secret Recovery Phase.
+						</div>
+					</div>
+				}
+			/>
+		</>
+	);
+};
+
+export default DeleteAllDataButton;
