+++ client/src/components/web-extension/Settings/index.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ContactIcon from '@cd/assets/image/contact-icon.svg';
 import LockIcon from '@cd/assets/image/lock-icon.svg';
 import { onClearPublicKey, lockAccount } from '@cd/actions/userActions';
+import DeleteAllDataButton from '@cd/components/web-extension/Common/DeleteAllDataButton';
 import SettingRow from './SettingRow';
 import './index.scss';
 
@@ -36,29 +37,22 @@ const SETTINGS = [
 			navigate('/welcomeBack');
 		},
 	},
-	{
-		name: 'Delete all data',
-		hasConfirmPopup: true,
-		icon: {
-			className: 'cd_we_setting_lock',
-			img: <LockIcon />,
-		},
-		action: ({ dispatch, navigate }) => {
-			dispatch(lockAccount());
-			navigate('/connectAccount');
-		},
-	},
 ];
 
 const Settings = () => {
 	return (
-		<section className="cd_we_single_section no_bottom_bar">
-			{SETTINGS.map((setting) => {
-				return (
-					<SettingRow key={setting.name} setting={setting} />
-				);
-			})}
-		</section>
+		<>
+			<section className="cd_we_single_section no_bottom_bar">
+				<div>
+					{SETTINGS.map((setting) => {
+						return <SettingRow key={setting.name} setting={setting} />;
+					})}
+				</div>
+				<div className="cd_setting_delete_btn">
+					<DeleteAllDataButton />
+				</div>
+			</section>
+		</>
 	);
 };
