+++ client/src/components/web-extension/Settings/index.scss
@@ -1,36 +1,52 @@
-.cd_setting_row {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	padding-bottom: 32px;
-	cursor: pointer;
-	.cd_setting_row_left {
+
+.cd_setting--wrapper {
+	.cd_setting_row {
 		display: flex;
+		justify-content: space-between;
 		align-items: center;
-		.cd_setting_name {
-			font-size: 16px;
+		cursor: pointer;
+	
+		&:hover {
+			color: var(--primary-light-color);
 		}
-		.cd_setting_icon {
+		&_left {
 			display: flex;
 			align-items: center;
-			justify-content: center;
-			margin-right: 10px;
-			width: 30px;
-			height: 30px;
-			border-radius: 50%;
-			&.cd_we_setting_about {
-				background: var(--primary-light-color);
+			.cd_setting_name {
+				font-size: 16px;
 			}
-			&.cd_we_setting_sites {
-				background: #b9ccfb;
-			}
-			&.cd_we_setting_lock {
-				background: #daf3cd;
+			.cd_setting_icon {
+				display: flex;
+				align-items: center;
+				justify-content: center;
+				margin-right: 10px;
+				width: 30px;
+				height: 30px;
+				border-radius: 50%;
+				&.cd_we_setting_about {
+					background: var(--primary-light-color);
+				}
+				&.cd_we_setting_sites {
+					background: #b9ccfb;
+				}
+				&.cd_we_setting_lock {
+					background: #daf3cd;
+				}
 			}
 		}
 	}
+
+	.cd_setting_modal {
+		text-align: center;
+	}
+	
+	&:not(:last-child) {
+		padding-bottom: 32px;
+	}
 }
 
-.cd_setting_modal {
-	text-align: center;
+.cd_setting_delete_btn {
+	margin-top: 74px;
+	display: flex;
+	justify-content: center;
 }
